+++ client/src/app/App.jsx
@@ -4,7 +4,7 @@ import Navbar from '../components/Navbar.jsx';
 import SongList from '../components/SongList.jsx';
 import TopAlbum from '../components/TopAlbum.jsx';
 import TopTenTracksByYear from '../components/TopTracksByYear.jsx';
-import DisplayYear from '../components/DisplayYear.jsx';
+// import DisplayYear from '../components/DisplayYear.jsx';
 import '../../styles/index.scss';
 
 export function App() {
@@ -18,6 +18,7 @@ export function App() {
       <SongList/>
       <TopAlbum/>
       </div>
+      <TopTenTracksByYear/>
     </>
   )
 }
