+++ client/styles/index.scss
@@ -11,6 +11,8 @@ $border-radius: 69px;
 html {
   width: 100%;
   height: 100%;
+  margin: 0;
+  padding: 0;
 }
 
 #root {
@@ -22,9 +24,10 @@ html {
 }
 
 body {
-  background: linear-gradient(45deg, #130188 0%, darken(#13008b, 50%) 100%);
-  background-size: 200% 200%;
-  height: 100%;
+  // background: linear-gradient(45deg, #130188 0%, darken(#13008b, 50%) 100%) bottom;
+  background: linear-gradient(to bottom,#130188, #13008b);
+  background-size: 400% 200%;
+  height: 100% !important;
   width: 100%;
   animation: background 20s ease infinite;
   display: flex;
@@ -218,7 +221,7 @@ ul {
 
 // This is the slider container
 
-.displayYearWrapper {
+.topTracksByYearWrapper {
   width: 100%;
   display: flex;
   flex-direction: column;
@@ -230,13 +233,23 @@ ul {
   display: flex;
   flex-direction: column;
   align-items: center;
+  margin-top: 100px;
+}
+
+.topTracksByYearYear {
+  background: $gradient-text;
+    -webkit-background-clip: text;
+  background-clip: text;
+    -webkit-text-fill-color: transparent;
+  font-size: 74px;
+  margin: 0px;
 }
 
 // This is the actual slider
 .slider {
   --webkit-appearance: none; // removing default css style
   appearance: none;
-  width: 100%;
+  width: 80%;
   height: 10px;
   margin: 50px 0px;
   border-radius: $border-radius;
